tests/submission/W3C/domhtml/HTMLIFrameElement07.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-iframe-element.html#dom-iframe-name" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe height="50" name="Iframe1" src="/support/html/sample.html" width="60">IFRAME1</iframe>


<script>
var domtest = async_test("The name IDL attribute of the frame element must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vname;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vname = testNode.name;

      assert_equals(vname,"Iframe1");
       
    domtest.done();
}));

</script>
<p>The name IDL attribute of the frame element must reflect the respective content attribute of the same name.</p>
<div id="log"></div>
</body></html>
